+++ src/main/java/duke/exception/DukeBlankDescriptionsException.java
@@ -1,4 +1,4 @@
-package duke;
+package duke.exception;
 
 public class DukeBlankDescriptionsException extends Exception {
 }
