+++ src/main/java/duke/exception/DukeInvalidTaskIndexException.java
@@ -1,4 +1,4 @@
-package duke;
+package duke.exception;
 
 public class DukeInvalidTaskIndexException extends Exception {
 }
