+++ src/main/java/duke/exception/DukeMissingDataException.java
@@ -1,4 +1,4 @@
-package duke;
+package duke.exception;
 
 public class DukeMissingDataException extends Exception {
 }
