+++ src/main/java/duke/task/Deadline.java
@@ -1,4 +1,6 @@
-package duke;
+package duke.task;
+
+import duke.task.Task;
 
 public class Deadline extends Task {
     protected String by;
