+++ src/main/java/duke/task/ToDo.java
@@ -1,4 +1,6 @@
-package duke;
+package duke.task;
+
+import duke.task.Task;
 
 public class ToDo extends Task {
     public ToDo(String description) {
